pyproject.toml
[tool.poetry]
name = "pdfriend"
version = "0.0.0"
description = "A command line utility for easily modifying PDF files"
authors = ["A. Grigoras"]
license = "GPLv3"
readme = "README.md"
scripts = { "pdfriend" = "pdfriend.main:main" }

[tool.poetry.dependencies]
python = "^3.10"
pypdf = "^3.16.1"
Pillow = "^10.0.1"
platformdirs = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
